+++ selfdrive/car/honda/values.py
@@ -87,7 +87,7 @@ class CAR:
   FIT = "HONDA FIT 2018"
   FREED = "HONDA FREED 2020"
   HRV = "HONDA HRV 2019"
-  HRV_2023_NON_US = "HONDA HRV/VEZEL 2023 NON US"
+  HRV_2023 = "HONDA HR-V 2023"
   ODYSSEY = "HONDA ODYSSEY 2018"
   ODYSSEY_CHN = "HONDA ODYSSEY CHN 2019"
   ACURA_RDX = "ACURA RDX 2018"
@@ -139,7 +139,7 @@ CAR_INFO: Dict[str, Optional[Union[HondaCarInfo, List[HondaCarInfo]]]] = {
   CAR.FIT: HondaCarInfo("Honda Fit 2018-20", min_steer_speed=12. * CV.MPH_TO_MS),
   CAR.FREED: HondaCarInfo("Honda Freed 2020", min_steer_speed=12. * CV.MPH_TO_MS),
   CAR.HRV: HondaCarInfo("Honda HR-V 2019-22", min_steer_speed=12. * CV.MPH_TO_MS),
-  CAR.HRV_2023_NON_US: HondaCarInfo("Honda HR-V/VEZEL 2023 NON US", "All", video_link="https://youtu.be/1O1u9jzpKD0", harness=Harness.bosch_b),
+  CAR.HRV_2023: HondaCarInfo("Honda HR-V 2023", "All", video_link="https://youtu.be/1O1u9jzpKD0", harness=Harness.bosch_b),
   CAR.ODYSSEY: HondaCarInfo("Honda Odyssey 2018-20"),
   CAR.ODYSSEY_CHN: None,  # Chinese version of Odyssey
   CAR.ACURA_RDX: HondaCarInfo("Acura RDX 2016-18", "AcuraWatch Plus", min_steer_speed=12. * CV.MPH_TO_MS),
@@ -1404,17 +1404,21 @@ FW_VERSIONS = {
       b'78109-THW-A110\x00\x00',
     ],
   },
-  CAR.HRV_2023_NON_US: {
+  CAR.HRV_2023: {
     (Ecu.eps, 0x18DA30F1, None): [
       b'39990-3M0-G110\x00\x00',
+      b'39990-3W0-A030\x00\x00',
     ],
     (Ecu.gateway, 0x18DAEFF1, None): [
       b'38897-3M0-M110\x00\x00',
+      b'38897-3W1-A010\x00\x00',
     ],
     (Ecu.srs, 0x18DA53F1, None): [
       b'77959-3M0-K840\x00\x00',
+      b'77959-3V0-A820\x00\x00',
     ],
     (Ecu.combinationMeter, 0x18DA60F1, None): [
+      b'78108-3V1-A220\x00\x00',
       b'78109-3M0-M310\x00\x00',
     ],
     (Ecu.fwdRadar, 0x18DAB0F1, None): [
@@ -1422,13 +1426,19 @@ FW_VERSIONS = {
     ],
     (Ecu.vsa, 0x18DA28F1, None): [
       b'57114-3M6-M010\x00\x00',
+      b'57114-3W0-A040\x00\x00',
     ],
     (Ecu.transmission, 0x18DA1EF1, None): [
+      b'28101-6EH-A010\x00\x00',
       b'28101-6JC-M310\x00\x00',
     ],
     (Ecu.programmedFuelInjection, 0x18DA10F1, None): [
+      b'37805-6CT-A710\x00\x00',
       b'37805-6HZ-M630\x00\x00',
     ],
+    (Ecu.electricBrakeBooster, 0x18DA2BF1, None): [
+      b'46114-3W0-A020\x00\x00',
+    ],    
   },
   CAR.ACURA_ILX: {
     (Ecu.gateway, 0x18daeff1, None): [
@@ -1548,7 +1558,7 @@ DBC = {
   CAR.FIT: dbc_dict('honda_fit_ex_2018_can_generated', 'acura_ilx_2016_nidec'),
   CAR.FREED: dbc_dict('honda_fit_ex_2018_can_generated', 'acura_ilx_2016_nidec'),
   CAR.HRV: dbc_dict('honda_fit_ex_2018_can_generated', 'acura_ilx_2016_nidec'),
-  CAR.HRV_2023_NON_US: dbc_dict('honda_civic_ex_2022_can_generated', None),
+  CAR.HRV_2023: dbc_dict('honda_civic_ex_2022_can_generated', None),
   CAR.ODYSSEY: dbc_dict('honda_odyssey_exl_2018_generated', 'acura_ilx_2016_nidec'),
   CAR.ODYSSEY_CHN: dbc_dict('honda_odyssey_extreme_edition_2018_china_can_generated', 'acura_ilx_2016_nidec'),
   CAR.PILOT: dbc_dict('acura_ilx_2016_can_generated', 'acura_ilx_2016_nidec'),
@@ -1568,6 +1578,6 @@ HONDA_NIDEC_ALT_PCM_ACCEL = {CAR.ODYSSEY}
 HONDA_NIDEC_ALT_SCM_MESSAGES = {CAR.ACURA_ILX, CAR.ACURA_RDX, CAR.CRV, CAR.CRV_EU, CAR.FIT, CAR.FREED, CAR.HRV, CAR.ODYSSEY_CHN,
                                 CAR.PILOT, CAR.RIDGELINE}
 HONDA_BOSCH = {CAR.ACCORD, CAR.ACCORDH, CAR.CIVIC_BOSCH, CAR.CIVIC_BOSCH_DIESEL, CAR.CRV_5G,
-               CAR.CRV_HYBRID, CAR.INSIGHT, CAR.ACURA_RDX_3G, CAR.HONDA_E, CAR.CIVIC_2022, CAR.HRV_2023_NON_US}
+               CAR.CRV_HYBRID, CAR.INSIGHT, CAR.ACURA_RDX_3G, CAR.HONDA_E, CAR.CIVIC_2022, CAR.HRV_2023}
 HONDA_BOSCH_ALT_BRAKE_SIGNAL = {CAR.ACCORD, CAR.CRV_5G, CAR.ACURA_RDX_3G}
-HONDA_BOSCH_RADARLESS = {CAR.CIVIC_2022, CAR.HRV_2023_NON_US}
+HONDA_BOSCH_RADARLESS = {CAR.CIVIC_2022, CAR.HRV_2023}